I2C: Make API asynchronous
All I2C block operations now works asynchronous if there is support
(ie: support to pthreads).

An important change here is that now the same I2C bus handle is shared
between users of the same bus, this way is easy to know if the bus is
busy with an operation on a different place.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>