Avoid re-definition of SOL_PLATFORM_LINUX
In sol-common-buildopts.h, we define SOL_PLATFORM_<$os> so it's available at
build time for users of the library. But during the build, all the config
values selected are available through the auto-generated sol_config.h, so by
naming the entry SOL_PLATFORM_LINUX we were effectively having it re-defined.

Changing the name in the config also keeps it consistent with the others.

Signed-off-by: Iván Briano <ivan.briano@intel.com>