conffile: Guarantee that 'file_reader' won't leak
Spotted by Coverity.

At a first glance, this looks like the reincarnation of '1cd044', but
with the new conffile code, whenever 'config_file_contents.len' is zero,
the inner function frees file_reader, so I think this time is a false
positive from Coverity.

However, the link between the functions is subtle and last time this
wasn't a false positive. It's a very simple check that won't happen very
often, so instead of flag the issue as 'false positive', I decided to
play safe and add the check back because the code isn't a big deal and
is more future proof.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>