mainloop: generic way to add an event source.
based on what GSource and GSourceFuncs, this is our way to allow
extending the main loop with other event sources, so we can have both
POSIX and GLIB main loops to be used with sol-bus.

I'm not making this a Linux-specific API since this is a generic
concept. However none of Contiki or RIOT uses prepare/check/dispatch,
will leave this out so the ports can adapt themselves to it.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com