common: Rename 'platform name' to 'board name'
Since is the board name that is expected and not the platform
name. This should avoid misinterpretations that were already happening.

The detection system was automatically setting the board name to
'riotos' when RIOT was found but it should be the name of where RIOT
is supposed to run (or empty if not known) that should be used e.g.
'intel-galileo-rev-d'.

Since there is no demand for a set of board related API's yet, instead
of change 'sol-platform-name-get' to 'sol-board-name-get' and have it
moved to it's on file and header file that would need to be installed, I
decided to name the function 'sol-platform-get-board-name' and keep it
on 'sol-platform' files.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>