mainloop: add sol-glib-integration.h
This header file is to be installed on linux systems, regardless of
glib being present or not. It's not compiled in libsoletta, so no
dependencies at compile time.

Users of glib (GMainLoop or GMainContext) should include this file and
call sol_glib_integration() to make sure glib will be usable with
libsoletta.

This works in all cases, it does not matter if libsoletta was compiled
with glib or posix main loop.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com