buffer: allow buffers using memory that is not owned by it.
This includes fixed capacity buffers or those that we don't want
free() to be called at sol_buffer_fini().

To achieve this in a nice way, rename 'reserved' to 'capacity', so
flag will match the structure member.

Fixed capacity buffers will not do realloc() as sol_buffer_resize() is
not allowed (-EPERM). This is useful when we have a buffer that is
allocated elsewhere, such as part of an UDP packet in CoAP, and you
don't want to allow reallocating it, but want to use the sol_buffer
functions to fill it.

We also do not allow realloc() on buffers that we can't free, as
realloc() implies in free() that buffer if there is no space.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com