buffer: allow fixed-sized buffers.
Fixed sized buffers will not do realloc() as sol_buffer_resize() is
not allowed (-EPERM). This is useful when we have a buffer that is
allocated elsewhere, such as part of an UDP packet in CoAP, and you
don't want to allow reallocating it, but want to use the sol_buffer
functions to fill it.

This also introduces a sol_buffer_fini_nofree(). It is basically
useless, but it's better to keep it for consistency instead of
documenting 'sol_buffer_fini() is not needed' since no free() should
be done.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com