test: improve test-mainloop-linux
The test-mainloop-linux does various checks for timeout and idle
executions, and expects to later receive a SIGTERM from its child, which
is handled and causes the mainloop to be interrupted.

Instead of using a usleep() to decide when to send SIGTERM, the child
will wait the parent to ask for it. Parent does it after all the other
checks are done. Child waits blocked on read() and when it's time,
parent closes the pipe.

Fixes one of the bots failing sometimes when running this test.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>