dep-resolver: add filesystem lookup
This patch introduces a new dependency check (namely filesystem), this
type of check will check the existence of a given list of files (attr:
files) in a list of possible paths (attr: path).

If found, dep-resolver will export a variable HAVE_{upper:dependency}
set to y and {upper:dependency}_PATH pointing to the path containing the
wanted files.

With that we can introduce the HAVE_ variable in the symbols dep chain
and use the _PATH variable to point to the proper file.

The first user for this infra-sctructure is the duktape thirdparty
module.

We're not actually using this resources in the build system, that
support will come in next patches.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>