build: support out-of-tree modules source code
This patch introduces the support for out-of-tree modules source code,
that is, a soletta's module that relies upon an external source code.

Duktape is the first user, but it's valid to any submodule. These
submodules may be checked out out of tree and maintained that way. The
build system will understand it's out-of-tree and adjust accordingly.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>