sol-iio: Support finding devices id by their relative path/name on sysfs
Introduce sol_iio_resolve_device_address(). This functions takes an
address and returns IIO device id. address can be of some forms:
starting with a '/' it points to absolute path of device on sysfs.
Currently, on the form of 'i2c/X-YYYY', for i2c device and X being the
bus number and YYYY is the device number, eg, i2c/7-0069.
Finally, if otherwise, will treat it as device name, and will return the
id of first i2c device with this name.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>