sol-iio: List of commands to find/create IIO device on sysfs
Introduce sol_iio_address_device(). This functions takes a list
of commands and returns, on a provided callbac, the IIO device id.
Commands can be of some forms:
starting with a '/' it points to absolute path of device on sysfs.
Currently, on the form of 'i2c/X-YYYY', for i2c device and X being the
bus number and YYYY is the device number, eg, i2c/7-0069.
Starting with 'create,' will attempt to create IIO device. Only I2C
creation is supported now.
Finally, if otherwise, will treat it as device name, if string,
and will return the id of first i2c device with this name. If a number,
will check if there's an IIO device with that id and return, on
callback, it's id or -1 in case there's no device with this id.
Commands will be executed from left to right stopping on first that
works.
Check sol_iio_address_device() docs to have a better explanation
of commands.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>