module string: Add USE_ICU
Add option USE_ICU and depends on it instead of HAVE_ICU, that way we
can enable and disable it regardless whether our system support. I was a
bit unsure where to put it, so I kept near the only user of it.

Configuration options like HAVE_* and FEATURE_* are defined by the
dependency resolver and the system. Ideally the code should consume a
different config option (conventionally USE_*) which allows disabling
something even if we HAVE it in the system.

Turns out that this is helpful for cross-builds, in which the
dependency-resolver doesn't cover all bases yet. We can forcefully
disable things until dep-resolver catch up.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>