flow: replace get_ports_counts callback with the counts
The node types always have the port counts when being created (or at
least initialized), so just provide those as attributes in the node_type
structure. This simplifies the flow code and reduces (a little) the
overall node type code.

Also finally removes the variables we've been keeping to track the ports
counts for each node -- when a per-type count would suffice.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>