fbp-runner: tweak the nodes used for simulation
Pushbutton goes down and up when clicked, so it's more useful for things
that are just a "signal" (empty packets). Use its PRESSED port to ensure
that one signal is sent.

Toggle button have two states (don't go up when first released, need to
press and release again) which matches better the boolean packet type.

Last but not least, label now knows how to give meaningful output to a
series of empty packets, so use it.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>