fbp-generator: clean up the information collection
Before actual generation, there's some information that refers to all
the FBP types being generated -- the loop was local to header
generation, but with this change we move it a bit outside. A further
patch will use it to collect types to initialize in the same pass.

This patch is only refactoring, and shouldn't change any behavior.

In the long run, I imagine this struct context moving out the generate
function to main, and the vector of FBP types be part of it (as well as
some other globals).

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>