fbp-generator: improve the -s option
By default, the generator produces a complete program, with
SOL_MAIN_DEFAULT() and a startup function.

However it's useful to have just a way to get the type itself, either
because the driving code (startup) will be provided by the user or
because there's no need for driving code, e.g. part of a module.

Add an argument to the existing '-s' option so that a symbol name must
be provided. It will be used to create a function that returns the
type (and does any internal initialization needed). The rest of the code
is just static implementation detail.

We may later improve this to generate a C header file if necessary, but
for now the users should use:

    extern const struct sol_flow_node_type *SYMBOL(void);

and compile/link the generated C code.

Last but not least, improve a bit the usage message.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>