fbp-generator: use sol_buffer
Instead of writing to a temporary file and then rename to the final
file, simply use a memory buffer to store the data -- now it is easy
because of sol_buffer_append_printf.

In some cases, temporary file was staying alive after the execution,
leading to garbage being kept around. This patch fixes this by not using
temporary files.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>