datatypes: add sol_vector_append_n()
Appends n elements to a vector and return a pointer to the first. This
is useful when the numbers of elements being added is know in advance
for two reasons: at most one allocation will happen and simplifies error
handling.

Returning a pointer to the first new element make it easy to iterate on
the added elements.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>