datatypes: add sol_vector_extend()
The function extends a vector by `amount` elements. This is useful when
we are going to append a known number of elements to a vector: at most
one allocation will happen and also simplifies error handling.

In the same spirit of `append`, `extend` returns a pointer to the first
new element, making easy to iterate in the newly added elements.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>