datatypes: make vector initialize memory to zero
Change sol_vector_extend() (and append) to initialize the memory to zero
before giving it back to the user. This is a saner default for the
typical case -- reducing the amount of potential garbage that the user
might accidentally deal with.

Only initialize memory up to len, so that memory read from outside the
len (but already allocated due to capacity) will still be flagged by
valgrind and similar tools.

The internal function grow() wasn't changed so that other functions that
do different kind of initialization (e.g. ptr_vector_insert_at()) don't
get more work than they need.

In the unlikely case in which this become a bottleneck for some
codepath, we may add a non-initializing variant of extend.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>