flow: decouple type creator functions from sol-flow-parser
Instead of passing internal parse_state struct, pass a context struct
that can become public. This is the first step towards moving JS (and
potentially other metatypes) outside sol-flow-parser.

A creator function should use the 'contents' (which is usually a
filename) to create a new type (and possibly other dependent types). For
each of those types, 'store_type()' callback must be called. The topmost
type should be returned in the output parameter.

The FBP creator function is a bit special, since it reuses the same
sol-flow-parser to parse the new buffer (potentially recursively), so it
will already have the types stored in the parser.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>