sol-iio: Read channel raw buffer
Add function to read a channel raw buffer - only when buffer is enabled.
Thus, it is possible to users read a channel bigger than 64 bits. Data
is returned via `sol_str_slice`. Note that this is a raw reading, so no
offset or scale is applyed. It's user task to parse this data.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>