shared: add sol_lib_loader
Library loader wraps the functionality of using dlopen() / dlsym() and
caching the results. It was based on looking at code from pin mux and
linux micro that does similar work. It will be used for loading flow
metatypes.

One feature that I kept out but might be useful in the future is to let
user associate arbitraty data to each library. It should be added once
there's a use case for it.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>