flow: add support for metatype modules
Use the "module" build infrastructure for finding metatypes to be used
when processing "DECLARE=..." statements in an FBP file.

These modules will be in modules/flow-metatype/ and they should follow
the interface defined by sol-flow-metatype.h, in particular use the
macro SOL_FLOW_METATYPE(). Currently the only job of such module is to
create a new type given a contents (usually a filename).

When running with dynamic libraries enabled, the flow parser will use
the library loader to look for modules that implement metatypes, so the
loaded metatypes are cached. As expected, if there are builtin types,
those are loaded too.

For now, both FBP and JS are handled "outside" this infrastructure. FBP
will remain special for a while, but JS will be soon moved to be a
module.

New types using those metatypes can also be created directly from C
code, by using sol_flow_parse_*_metatype() functions. This reduces the
need of a "per metatype" C API. This is implemented providing to the
metatype a "read_file" a special read_file function that can read only
one file, the one provided by the buffer.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>