build: add Requires.private to soletta.pc
This patch introduces a new private-requires-y and private-requires-m
variables to declare which pkg dependencies we're using, with that we
can add that to our distributed soletta.pc file.

For that we're also considering env vars in the template parser so we
can easily export&get useful information while processing template
files.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>