Fix make samples
Issue introduced in: Add a trash disposer flow application

There was 2 problems on this commit:
- the json configuration file was not set in Makefile, so fbp don't know
the type of the nodes defined on configuration file.

- the servo node is not a builtin node we need correctly handle this case in
fbp generator also when building the generated file, this issue will be
handled in future so for now it only will build trash-disposer if servo node
was set as builtin by user.

Signed-off-by: José Roberto de Souza <jose.souza@intel.com>