build: split static and shared lib build
This patch changes how shared and static libraries are built. Currently
we build both shared and static, with this patch we'll build only one of
them. The user can choose to build static or library using kconfig
interfaces, the default being shared.

The main motivation for this split was the compilation with some
embedded platforms using the static library showed problematic due to
the indiscriminate use of -fPIC for both shared and static library. We
faced memory addresses issues with global non static uninitialized
variables - this may result in more problems than that, we just didn't
reach them out.

If static is chosen the .so external modules are not built, everything
goes builtin - actually it's currently done that way, but since we were
building the shared library together we should also generate the
external modules so it could be used by the shared library users.

Next step - as requested by Jesus - is to split the flags tested by
dependency-resolver, so we don't introduce compilation flags specific to
shared libraries and get the static build affected.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>