Remove old tests-fbp directory
Follow our current convention:

- "Tests" are focused on ensuring that the functionality works, and must
  run automatically (or have a description of steps to run), and a way
  to verify whether they work. E.g. src/test/, src/test-fbp/,
  src/test-stubgen.

- "Samples" are focused on showing how Soletta can be used, they contain
  comments / explanations to help understanding how things
  work. E.g. src/samples/.

Of course we want the samples to be always working (there will be a
procedure for the QA to test them), and the tests to not be
unreadable. But there is a difference in what they accomplish. A sample
doesn't need to address ALL the options available (in the same sample)
of a feature, while this kind of coverage is welcome in the tests.

The two removed files didn't fit either category (no comments, no
instructions, neither automatic ran). For those one-shot development
tests of a node type that doesn't deserve an automatic test, we now have
a way to use sol-fbp-runner to load a node type with options

      sol-fbp-runner -t timer -o interval=3000

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>