comms: Turn sol-socket into a very simple BSD sockets wrapper
In order to port the functionality to other platforms that don't have
BSD sockets, we need to wrap them on our own API first. And since
sockets can be used on a wide variety of things, we need to start small
and only with that we need. For now, wrap only the required
functionality to make CoAP work.

Signed-off-by: Iván Briano <ivan.briano@intel.com>