flow: remove location as option
Location isn't handled as option. It's only a packet type.
We don't have support or use cases to use it as option, so
removing it from schema and generator.

If at some point we decide to support it as option, we
can add it back on schema and generator and fully
implement support to that.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>