modules/console: clean up process function
Create a function to provide the output with prefix / suffix and use it
in all cases that is possible. Two cases had special output that
wouldn't fit in a single formatting and were kept as is.

Prefix always contains a string and suffix was changed to be the same
way, this avoid all those checks for NULL everywhere.

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>