flow/random: Store state as `unsigned int` to avoid UB in 32-bit
This has been reported on Issue #607, where undefined behavior
sanitizer would produce runtime errors when generating random numbers
on a 32-bit machine.

Move state to `unsigned int` so that the overflow behavior is defined,
and, where signed data is expected, shift value right by 1 bit before
casting to an integer in order to kill the bit that would be the sign
bit.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>