flow/random: Initialize random seed with platform-specific methods
If no seed is specified, use a platform-specific entropy pool to
initialize the PRNG node type.

Under Linux, try using getrandom(2), falling back to reading from
/dev/urandom.

On other platforms (or, in the case that both methods for Linux), fall
back to using a bad, but frequently found source of entropy:
time(NULL).  Since the random node type is not expected to be used for
anything other than a simple PRNG, this won't be a problem.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>