build: properly unquote CONFIG_CFLAGS and CONFIG_LDFLAGS
When using CONFIG_CFLAGS containing spaces, it would be appended to the
command line keeping the quotes. The issue is that `patsusbt` operates
on words, but not assuming quotes mean anything.

    $(patsubst "%",%,$(CONFIG_CFLAGS))

was evaluating

    "-Wno-undef
    -Wno-missing-include-dirs"

and neither matched.

Using two patsubst, one for "% and another for %" works in some cases
but would break in the case we had quotes inside one of the flags.

This patch uses `echo`, and ensures that it is called only once for each
flag by creating the UNQUOTED_ variables using `:=` (single assignment).

Signed-off-by: Caio Marcelo de Oliveira Filho <caio.oliveira@intel.com>