Add new UUID nodes and linux-micro service.
The string/uuid node outputs fresh, new v4 UUIDs on a flow.

The service (which is not actually a service, but a run-once-only
routine) will run after fstab one (if it's present, otherwise it tries
without it anyway), and will guarantee that a machine-id exists (in
known places) in the file system.

The string/machine-id node gets said machine-id and outputs it on a
flow.

Tests for the nodes are included.