dep-resolver: let the users overwrite project's definitions
With this change now users may overwrite definitions values, i.e for
distros building systems we may want to intruct the LIBDIR based on
target platform i.e /usr/lib64 for x86-64. For that the user can for
example export LIBDIR=/usr/lib64/ and the the $(build_libdir) will be
properly be adjusted.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>