modules: just warn and change wrong options to default on buffer
int/buffer and boolean/buffer are failing to open if an option is wrong.
We're following the design of using default option in this case
and warn the developer about it.

Also, 'samples' check should be for <= 0, not only == 0.
Otherwise it will try to work with negative ammount of samples.

Signed-off-by: Bruno Dilly <bruno.dilly@intel.com>