pin-mux: modules refactoring
We came to the conclusion that pin mux modules should consist of just a set
of entry points on which board specific code could be executed in order to
do the work.

This way, all multiplexing logic stays inside the modules allowing our
infrastructure to handle different approaches of multiplexing that may
be present in different boards. Instead of only handle GPIO controlled
multiplexers.

This setup also allowed us to move to pin mux modules some pieces of code
dealing with board specific details that were inside Soletta, like
Galileo's 'drive' and Edison's mode setup (that will be added next).

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>