shared: Introduce sol-rng, a Random Number Generator API
This is a simple abstraction to generate random numbers. It is
loosely-based on the C++11 API supported by <random> include, sans the
distribution stuff.

Current implementation contains three engine implementations: a MT19937
(default), a random_r() generator (if available; currently only glibc
supports it), and a /dev/urandom reader.

Only raw bytes can be generated by the API at the moment.  The idea is
to implement custom type generation (with proper distribution) as the
need arises.

Currently, the implementation is completely opaque, but can be opened
up if the need to let Soletta API users implement their own engines
arise.

Usage is pretty simple at the moment:

	struct sol_rng_engine *engine;
	uint64_t value;

	/* Passing the default engine or an engine not supported by
	 * the platform will fallback to the MT19937 engine. */
	engine = sol_rng_engine_new(SOL_RNG_ENGINE_IMPL_DEFAULT);

	/* Passing 0 to the seed will try to use a good source of
	 * entropy available in the system.  This is /dev/urandom on
	 * Linux (using getrandom() if available). */
	sol_rng_engine_seed(engine, 0);

	assert(sizeof(value) == sol_rng_engine_generate_bytes(engine, &value, sizeof(value)));

	sol_rng_engine_destroy(engine);

As a bonus, move random node types to use this new API.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>