board-detect: split from linux and allow optional detection.
Move board detect check to be independent from linux, just depend on
FEATURE_FILESYSTEM at this moment (that only Linux implements).

The board-detect is then using its own Kconfig option and can be
explicitly disabled in the case we're compiling to a single board.

Also take the time to always validate the board name, not only when it
came from envvar.

Signed-off-by: Gustavo Sverzut Barbieri gustavo.barbieri@intel.com