io: New persistence storage, memmap.
Useful for storing data on media that only provides a block of memory to
store on, such as EEPROM or NVRAM. To define how this memory should be
used, one must provide a map, using `sol_memmap_add_map()`. This map
contains size of data being saved, offset in memory and it's name.
Each map also has a `path` property, used to find the file representing
underlying storage. So it's possible to store on different storage
media.
On fbp files, node 'persistence' supports it when using 'memmap' on
storage option.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>