build: fix initial-service file name
In the commit 7393eef4c866324001c6cbf6e123ded2a5e2770a I re introduced
the initial-services file generation and installation, however, the dest
file name was misspelled and was missing an 's', it should be
initial-services instead of initial-service. This patch fixes the issue.

Also handles the case when all linux-micro mods are <m>.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>