comms:http: Ugly hack to make async response works
Suspend/Resume a connection using an external mainloop is not working in
the last release of libmicrohttpd. This patch is a workaround while
there is not a oficial release containing the revision r36290.

It should be reverted when the next libmicrohttpd release happened and
the properly version should be set as minimum dependency.

Signed-off-by: Flavio Ceolin <flavio.ceolin@intel.com>