run-fbp-tests: random changes and improvements
This patch changes a bit the way the run-fbp-tests processes the tests
to be run, these changes are meant to make things easier to add new skip
codes, parsing and so on.

Also new operations were added, --query-only will run the the tests dir
and determine the tests should be run, --skip defines which skip codes
are considered, --compiled runs the compiled tests instead of running
the fbp's with sol-fbp-runner, --compiled-dir is used to determine where
the tests reside.

The future patches will introduce the required changes to build system
so we can make use of it all.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>