efivars-storage: Fix path check function used
An efivar is actuallya file under `/sys/firmware/efi/efivars/`. When
user gives a name to a var, s[he] could use '../../' trickery to make
Soletta open a file on the wrong directory. We use `realpath()` to check
this, however, it only works for files which already exist. Currently,
creating a new variable will fail when `realpath()` fails. The solution
was to move `realpath()` check after file opening. We still refuse to
write to that file if it fails check, but a file can be created
nevertheless. We still need a better approach to this - like a path
sanitiser.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>