build: link to static libasan
In the commit 202bff12f21b114d39b818b887ff94771a034c4d we changed the
sanitze dependencies to also force linking against libasan - in the
commit message we state that this would be harmless to newer gcc
versions. I turns out not to be true, if we're linking against the
shared libasan we should make sure to preload the lib or make sure it is
the first DSO name[1].

To avoid this kind of runtime issues we're forcing libasan to be
statically linked.

[1] -
https://github.com/gcc-mirror/gcc/blob/master/libsanitizer/asan/asan_linux.cc#L112

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>