doxygen: generate doxygen files in build/stage dir
This patch first intends to change the doxygen files generation to
inside the build/ dir. It also changes the rules and dependencies chain,
now it only generate the files if required - considering as direct
dependency the doxygen configuration and templates.

The generation for all docs will be moved to build/stage/doc/

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>