*WON'T BE INTEGRATED - TESTING ONLY* A commit to test memmap
Not only memmap, but if one wants to see C API in use, this commit is
for that. I'll be working with @anselmolsm to find a way to put these
tests in use.
To use the map, one needs to pass '-m map.json' to sol-fbp-runner (or
generator).
To see the test, just run persistence.fbp. Before actually running it,
it should perform some batch tests on filesystem, efivars and nvram (
don't worry, it should write to a harmless 'nvram.bin' file, that must
exist prior running the test [and this explanation gives a glimpse why
these tests are not 'official' yet, these issues must be addressed on
our CI tools also]).

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>