build: make it possible to reuse objects inter-modules
With this patch now we can have common objects to be shared between
modules, currently we can do so by adding objects to core or shared and
make it widely exposed to all libsoletta and modules.

As a side effect we removed the need of two nested iterations to
generate the objects rules by just iterating the objects list instead of
modules + objects and buitins + objects.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>