common: Module loading helper
It allows users to have access to symbols regardless of whether they
are built into the library or if a module needs to be loaded for them.

The function sol_modules_get_symbol() will search for the requested
symbol in the library first, and if not found, it tries to load the
module of the corresponding namespace (both given).

Said function is not intended to be used directly, instead, the macro
sol_symbol_get() should be used, that resolves to the address of the
(unquoted) symbol passed to it in the case where dynamic modules
support is not enabled.

Signed-off-by: Iván Briano <ivan.briano@intel.com>