samples: Remove pointless check
Back when there was a single builtins.h it made sense to check for this
define value to see if a node type had been built. Now that the
specific module header is used, it either exists or it doesn't and the
define makes no sense anymore.

We could use the value as defined by Kconfig, but I'd rather leave
samples clean of build system cruft.

Signed-off-by: Iván Briano <ivan.briano@intel.com>