sol-bus: Add support for monitoring interfaces
This will allow users to be notified when an interface is added/removed
from a hierarchy.

Much of the code used for monitoring properties is used, we just need to
be careful and check the interface before calling the functions that
will handle changes in properties.

This is only possible because the content of the PropertyChanged,
InterfacesAdded signals and the response of the GetManagedObjects method
are consistent.

Signed-off-by: Vinicius Costa Gomes <vinicius.gomes@intel.com>