flow: Add string tests for the ascii backend
The current set of tests are all dependent on a build with icu to work.
So, when lacking this dependence, almost all string tests will fail
because of the string format instead of a miss behavior of our ascii
backend.

This patch add the ascii version of those tests so we can exercise the
ascii backend properly (what can be done by compiling soletta without
icu library).

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>