Flow: Fix sol_flow_packet_new() crash for String packets.
Soletta would crash of one tries to create a String packet using the
following code:

/* Code.... */

packet = sol_flow_packet_new(SOL_FLOW_PACKET_TYPE_STRING, "My String");

/* Code.... */

However it would not crash if one creates a String packet using the
following code:

/* Another code */

packet = sol_flow_packet_new_string("The String");

/* Another code */

The crash is happening because, string_packet_init()
is expecting a sol_str_slice as an input. The function
sol_flow_packet_new_string() creates the slice for the user
and then calls sol_flow_packet_new().

The API for creating packets must be consistent, the
user should be able to create a String packet using the
first code snippet. This patch changes string_packet_init()
in a way that it expects a plain string and not a
sol_str_slice.

This patch also fix another inconsistency, as a pointer to a string
must be provided to the sol_flow_packet_new().

Signed-off-by: Guilherme Iscaro <guilherme.iscaro@intel.com>