flow: Fix string corruption in sol-conffile.c
According to the manual, 'dirname' and 'basename' may change the input
string and 'sol_argv' return was being provided directly to them which
in turn was corruption it's internal string since it doesn't return a
copy.

Also, each call needs it's own copy to guarantee that none will receive
corrupted that as input.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>