flow/file: Fix segmentation fault while sending error packet
Functions taking a printf-style formatting string should never be
allowed to receive user-supplied strings.  In this case, a string has
been sent to `sol_flow_send_error_packet()` after being formatted by
`snprintf()` -- with `%s` and all -- causing the program to crash.

Modify the prototype for `sol_flow_send_error_packet()` so that it's
declared with `__attribute__((printf))` and fix all problematic places
in the tree with help from the compiler.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>